Add files via upload
The basic motive behind the neural network and classification is to identify patterns in language use that are characteristic of individuals with dementia. The neural network uses the transcriptions of conversations between individuals with dementia and their caregivers to learn these patterns and classify individuals as either having dementia or not. By accurately classifying individuals with dementia, the model can serve as an early diagnostic tool, allowing for earlier intervention and improved outcomes for patients.